Remove PDB reader
