--- src/Microsoft.Diagnostics.Runtime/src/Utilities/PDB/PdbException.cs
@@ -1,16 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.IO;
-
-namespace Microsoft.Diagnostics.Runtime.Utilities.Pdb
-{
-    internal class PdbException : IOException
-    {
-        internal PdbException(string format, params object[] args)
-            : base(string.Format(format, args))
-        {
-        }
-    }
-}
